Implement Production Settings & Memory Optimizations Implements the settings display as intended for production use. Also performs some memory optimizations to allow this code to be run on Arduino Uno Rev3. As a result, the amplitude and clipping history used by the auto gain has been truncated to 32 values (~2.1s).

+++ software/libraries/DMXFixture/DMXFixture.h
@@ -22,13 +22,13 @@ public:
     void display(DMX_Master &);
 
 private:
-    uint16_t _startChannel;
-    uint16_t _redChannel;
-    uint16_t _greenChannel;
-    uint16_t _blueChannel;
-    uint16_t _whiteChannel;
-    uint16_t _dimmerChannel;
-    uint16_t _strobeChannel;
+    uint8_t _startChannel;
+    uint8_t _redChannel;
+    uint8_t _greenChannel;
+    uint8_t _blueChannel;
+    uint8_t _whiteChannel;
+    uint8_t _dimmerChannel;
+    uint8_t _strobeChannel;
     uint8_t _dimmerDefaultValue;
     uint8_t _dimmerValue;
     uint8_t _rgbDimmerValue;
